+++ CHANGES.md
@@ -3,6 +3,7 @@ See also https://pvp.haskell.org/faq
 ## Version 1.4.0.0
 
  * `Eq` is now a superclass of `Hashable`.
+ * Remove `Hashable1 Fixed` instance
 
 ## Version 1.3.3.0